requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
PyPDF2>=3.0.0
pdfplumber>=0.11.0
pdf2image>=1.16.0
pytesseract>=0.3.10
opencv-python>=4.9.0
Pillow>=10.0.0
torch>=2.3.0
transformers>=4.44.0
sentencepiece>=0.2.0
evaluate>=0.4.0
pyyaml>=6.0.0
typer>=0.12.0
rich>=13.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

